Move to nemo-curator's logger
